+++ src/pages/reports/owners/summary/index.tsx
@@ -36,7 +36,7 @@ export default function OwnersReportSummary() {
   return (
     <div className="w-full flex flex-col gap-10">
       <OwnersReportTabSwitch state="summary" />
-      <div className="w-full grid grid-cols-1 md:grid-cols-3 gap-5">
+      <div className="w-full grid grid-cols-1 lg:grid-cols-3 gap-5">
         {[
           {
             id: 1,