+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 10.04.2025
+
+> Feature: Assign apartment to owner user
+> Completion of owners reports API integrations
+
 ##### 09.04.2025
 
 > Fix: Owners reports integration bugs resolution